Split Apps tests into separate files
